requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
fpdf2
gTTS
APScheduler
flask
flask-cors
SpeechRecognition
pydub
